First commit
A piece of software that helps managinc the export, import and delete actions of an xbox, ps4 or nsw console. This is WIP and cannot be treated as a fully usable product until all platforms are supported. This is for personal use only and does not contain all the external components that would make the program work as intended.